# src/express-server/api/openapi.yaml
openapi: 3.0.0
info:
  title: Food Delivery API
  version: 1.0.0
  description: API REST para sistema de delivery - Trabalho DW1 UMAIA 2025/26
  contact:
    name: inf25dw1g13
    email: a046756@umaia.pt
components:
  schemas:
...
    Pedido:
      title: Pedido
      type: object
      required:
        - cliente_id
        - restaurante_id
        - morada_entrega_id
        - metodo_pagamento
        - pratos
      additionalProperties: false
      properties:
        id:
          type: integer
          readOnly: true
          description: ID do pedido (gerado automaticamente)
        cliente_id:
          type: integer
          description: ID do cliente que fez o pedido
          example: 1
        restaurante_id:
          type: integer
          description: ID do restaurante
          example: 1
        morada_entrega_id:
          type: integer
          description: ID da morada de entrega
          example: 1
        metodo_pagamento:
          type: string
          enum: [cartao, mbway, dinheiro, carteira]
          description: "Método de pagamento"
          example: "mbway"
        pratos:
          type: array
          minItems: 1
          description: "Array de pratos do pedido (obrigatório)"
          items:
            type: object
            required:
              - prato_id
            properties:
              prato_id:
                type: integer
                description: ID do prato
                example: 1
              quantidade:
                type: integer
                minimum: 1
                default: 1
                description: "Quantidade do prato (padrão: 1)"
                example: 2
              observacoes_item:
                type: string
                description: "(Opcional) Observações específicas deste item"
                example: "Sem cebola"
        total:
          type: number
          format: float
          readOnly: true
          description: "Total do pedido em euros (calculado automaticamente)"
          example: 27.50
        estado:
          type: string
          enum: [pendente, confirmado, a_preparar, entregue, cancelado]
          readOnly: true
          description: "Estado do pedido (padrão: pendente)"
          example: "pendente"
        observacoes:
          type: string
          description: "(Opcional) Observações gerais do pedido"
          example: "Entregar na porta"
...
  examples:
...
    PedidoExample01:
      value:
        cliente_id: 1
        restaurante_id: 1
        morada_entrega_id: 1
        metodo_pagamento: "mbway"
        pratos:
          - prato_id: 1
            quantidade: 2
            observacoes_item: "Sem cebola"
          - prato_id: 2
            quantidade: 1
        observacoes: "Entregar na porta"
    PedidoExample02:
      value:
        cliente_id: 2
        restaurante_id: 2
        morada_entrega_id: 2
        metodo_pagamento: "cartao"
        pratos:
          - prato_id: 6
            quantidade: 1
          - prato_id: 7
            quantidade: 2
            observacoes_item: "Bem passado"
        observacoes: "Entregar na porta"
...
paths:
...
  /api/pedidos:
    get:
      tags: [Pedidos]
      x-eov-operation-handler: controllers/PedidosController
      operationId: getPedidos
      summary: Listar pedidos
      parameters:
        - name: estado
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Lista de pedidos
    post:
      tags: [Pedidos]
      x-eov-operation-handler: controllers/PedidosController
      operationId: createPedido
      summary: Criar pedido
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pedido'
            examples:
              PedidoExample01:
                $ref: '#/components/examples/PedidoExample01'
              PedidoExample02:
                $ref: '#/components/examples/PedidoExample02'
      responses:
        '201':
          description: Pedido criado
  /api/pedidos/{id}:
    get:
      tags: [Pedidos]
      x-eov-operation-handler: controllers/PedidosController
      operationId: getPedido
      summary: Obter pedido por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
          examples:
            one:
              value: 1
              summary: Retrieve Pedido id 1
            two:
              value: 2
              summary: Retrieve Pedido id 2
      responses:
        '200':
          description: Pedido encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pedido'
    put:
      tags: [Pedidos]
      x-eov-operation-handler: controllers/PedidosController
      operationId: updatePedido
      summary: Atualizar pedido
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
          examples:
            one:
              value: 1
              summary: Update Pedido id 1
            two:
              value: 2
              summary: Update Pedido id 2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pedido'
            examples:
              PedidoExample01:
                $ref: '#/components/examples/PedidoExample01'
              PedidoExample02:
                $ref: '#/components/examples/PedidoExample02'
      responses:
        '200':
          description: Atualizado
    delete:
      tags: [Pedidos]
      x-eov-operation-handler: controllers/PedidosController
      operationId: deletePedido
      summary: Apagar pedido
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
          examples:
            one:
              value: 1
              summary: Delete Pedido id 1
            two:
              value: 2
              summary: Delete Pedido id 2
      responses:
        '204':
          description: Removido